Fix to prevent RKRequestQueue from sending duplicates
The requestQueue used to check RKRequest's isLoading flag to determine
if a request needs to be sent. But it's turned off upon receiving a
response, some more work is done and then it's removed from the queue.
During "some more work" another thread can trigger the queue to send
"unsent" requests. This resulted in the loadingCounter incrementing
twice for the same request but only decrementing once. Eventually the
loadingCounter reaches our limit and the queue hangs.

This patch sets a new RKRequest property "isSentFromRequestQueue" when
the request is sent and uses that in the test for what to send. Also
the decrement counter statement is moved into a synchronized block
for safety.

This was rather difficult to debug due to the "noise" resulting from
the other places RKLog___ gets used and from the complexity of the
execution path so I left a _sentRequests dictionary and commented
NSLogs in place to make it easier to come back and re-analyze.

Fixes: https://litl.fogbugz.com/default.asp?35092